% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPredat.R
\name{PrPred}
\alias{PrPred}
\title{A function to take the parameters from the user and plot the Predator Abundance Vs Prey Abundance Plot.}
\usage{
PrPred(a, b, g, d)
}
\arguments{
\item{love}{Do you love plots? Defaults to TRUE.}
}
\description{
This function allows you to see different plots with different values of the set paremeters.
}
\examples{
PrPred()
}
\keyword{Predator}
\keyword{Prey}
\keyword{model}
\keyword{vs}

